// ANGULAR/ANGULAR/src/app/Shared/Enum/Common.enum.ts
export enum DepartmentCategoryEnum {
  Category = 1
}

export enum LookupEnum {
  Shilanyas = '30093',
  Lokarpan = '30094',
  UrbanDemoValue = '20069',
  UrbanProdValue = '20068',
  RuralDemoValue = '20070',
  RuralProdValue = '20067'
}

export enum LookupDataTypeEnum {
  DATA = 'DATA',
  PDF = 'PDF',
  URL = 'URL',
}

export enum ManualTypeEnum {
  UserManual_Val = 1,
  UserManual_Txt = 'User Manual',

  SystemManual_Val = 2,
  SystemManual_Txt = 'System Manual',
}

export enum UserSSOIdEnum {
  RajeshSaini = 'rajeshsaini.risl',
  CMISNewTest3 = 'cmisnewtest3',
}